[feat] get matches of an alias (untested).

+++ Models/DTO/LobbyDTO.cs
@@ -4,8 +4,8 @@ public class LobbyDTO
 {
     public int Id { get; set; } = -1;
     public long LobbyId { get; set; }
-    public long BestOf { get; set; }
-    public long Warmups { get; set; }
+    public int BestOf { get; set; }
+    public int Warmups { get; set; }
 
     internal LobbyDTO(Lobby? lobby)
     {
